+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 # docker-otrs Change Log
 
+## 6.0.9 - 2018-07-24
+### Changed
+- Updated to latest OTRS version 6.0.9.
+- improvements in custom skins handling.
+### Added
+- New environment variable OTRS_TIMEZONE to set the default timezone.
+
+## 6.0.8 - 2018-06-12
+### Changed
+- Updated to latest OTRS version 6.0.8.
+
+## 5.0.28 - 2018-06-12
+### Changed
+- Updated otrs-5_0_x branch to OTRS 5.0.28.
+
+## 4.0.30 - 2018-06-12
+### Changed
+- Updated otrs-4_0_x branch to OTRS 4.0.30.
+
+
 ## 6.0.7 - 2018-05-29
-### Added(
+### Added
 - Check for backup file integrity before starting the restore backup process (OTRS_INSTALL=restore).
 
 ### Changed